+ add OnDemandWorkloadService to avoid flushing collectors after each invoke.
