+++ chassis/core/src/main/java/com/griddynamics/jagger/engine/e1/process/PeriodWorkloadProcess.java
@@ -1,12 +1,15 @@
 package com.griddynamics.jagger.engine.e1.process;
 
+import com.google.common.util.concurrent.Service;
 import com.griddynamics.jagger.coordinator.NodeContext;
 import com.griddynamics.jagger.engine.e1.scenario.WorkloadConfiguration;
+import com.griddynamics.jagger.util.Futures;
 import com.griddynamics.jagger.util.TimeoutsConfiguration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Collection;
+import java.util.concurrent.Future;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
@@ -44,18 +47,45 @@ public class PeriodWorkloadProcess extends AbstractWorkloadProcess {
         }, delay, period, TimeUnit.MILLISECONDS);
     }
 
+    @Override
+    protected void startNewThread() {
+
+        for (WorkloadService thread : threads) {
+            Future<Service.State> future = thread.start();
+            if (future != null) {
+                Service.State state = Futures.get(future, timeoutsConfiguration.getWorkloadStartTimeout());
+                log.debug("Workload thread with is started with state {}", state);
+                return;
+            }
+        }
+
+        if (executor.getActiveCount() >= executor.getMaximumPoolSize()) {
+            log.warn("Thread pool(size={}) is full. Skip adding new thread.", executor.getPoolSize());
+            return;
+        }
+
+        super.startNewThread();
+    }
 
     @Override
-    protected void stopBeforeTerminating() {
-        // do not stat new workload service
-        loopExecutor.clear();
-        loopExecutor.shutdown();
+    protected WorkloadService getService(AbstractWorkloadService.WorkloadServiceBuilder serviceBuilder) {
+        // return workload service that should execute 1 sample on demand
+        return serviceBuilder.buildInvokeOnDemandWorkloadService();
     }
 
 
     @Override
-    protected void changeConfigurationAfterStats(WorkloadConfiguration configuration) {
-        // start new period schedule according to new configuration
+    public void stop() {
+        // do not stat new workload service
+        loopExecutor.clear();
+        loopExecutor.shutdown();
+
+        super.stop();
+    }
+
+
+    @Override
+    public void changeConfiguration(WorkloadConfiguration configuration) {
         loopExecutor.scheduleAtFixedRate(new Runnable() {
             @Override
             public void run() {
@@ -64,15 +94,5 @@ public class PeriodWorkloadProcess extends AbstractWorkloadProcess {
         }, configuration.getDelay(), configuration.getPeriod(), TimeUnit.MILLISECONDS);
     }
 
-    @Override
-    protected void changeConfigurationBeforeStats(WorkloadConfiguration configuration) {
-        // stop scheduling current task
-        loopExecutor.clear();
-    }
 
-    @Override
-    protected WorkloadService getService(WorkloadService.WorkloadServiceBuilder serviceBuilder) {
-        // return workload service that should execute 1 sample
-        return serviceBuilder.buildServiceWithPredefinedSamples(1);
-    }
 }
